feat: TestUtil step2 completed registation t4()_b
