+++ src/test/java/com/programmers/AppTest.java
@@ -1,11 +1,47 @@
 package com.programmers;
 
 import com.programmers.standard.util.TestUtil;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 import java.io.ByteArrayOutputStream;
 import java.util.Scanner;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
 public class AppTest {
+    @Test
+    @DisplayName("== 명언 앱 ==")
+    public void t1() {
+        String output = AppTest.run("");
+
+        assertThat(output)
+                .contains("== 명언 앱 ==");
+    }
+
+    @Test
+    @DisplayName("명령) ")
+    public void t2() {
+        String output = AppTest.run("""
+                목록
+                """);
+
+        assertThat(output)
+                .contains("명령) ");
+    }
+
+    @Test
+    @DisplayName("명령을 2번 이상 입력할 수 있습니다.")
+    public void t3() {
+        String output = AppTest.run("""
+                목록
+                목록
+                """);
+
+        String[] split = output.split("명령\\)"); //'명령) '을 기준으로 split(나누면) 그룹이 4개가 생김.
+        assertThat(split).hasSize(4);
+    }
+
     public static String run(String input) {
         input = input.stripIndent().trim() + "\n종료"; //종료 명령어 중복 제거를 위해 input에 따로 붙여주기
